Add the save_colorscheme.vim script
This script can be used to save the current colors.
To use, open a new buffer and `:source save_colorscheme.vim`.

If `g:save_all_highlights` is set to 1 before sourcing the script, all
highlight colors known by vim are listed, otherwise save the one listed
in `s:names` defined in the script.

Change the line with `g:colors_name` to give a reasonable name.